Handle holidays, calculate discounts, added tests (required and additional)

+++ src/main/java/com/toolshop/rental_system/model/ToolTypeInfo.java
@@ -1,12 +1,13 @@
 package com.toolshop.rental_system.model;
 
+import java.math.BigDecimal;
+
 import com.toolshop.rental_system.enums.ToolType;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -22,9 +23,7 @@ public class ToolTypeInfo {
     private Long id;
     @Enumerated(EnumType.STRING)
     private ToolType type;
-    // TODO ensure that double is acurate enough and that this doesn't need to be a
-    // float instead
-    private double dailyChargeAmount;
+    private BigDecimal dailyChargeAmount;
     private boolean weekdayCharge;
     private boolean weekendCharge;
     private boolean holidayCharge;
